Module + Map Update
Compressed all modules with moving parts into entities.py. My classes and functions intertwine with each other so we can't have them in separate modules.

Also added spawners and map for the first floor
