+++ resources/enemy_spawner.png
