+++ scripts/assets.py
@@ -3,7 +3,7 @@ import pyglet
 pyglet.resource.path = ['../resources']
 pyglet.resource.reindex()
 
-background_img = pyglet.resource.image('background.png')
+firstFloor_img = pyglet.resource.image('first_floor.png')
 playerF_img = pyglet.resource.image('player_Front.png')
 playerB_img = pyglet.resource.image('player_Back.png')
 playerL_img = pyglet.resource.image('player_Left.png')
@@ -12,12 +12,13 @@ obstacle_img = pyglet.resource.image('obstacle.png')
 monsterSmall_img = pyglet.resource.image('monster_small.png')
 monsterMedium_img = pyglet.resource.image('monster_medium.png')
 monsterLarge_img = pyglet.resource.image('monster_large.png')
+spawner_img = pyglet.resource.image('enemy_spawner.png')
 
 def center_image(image):
     image.anchor_x = image.width//2
     image.anchor_y = image.height//2
 
-center_image(background_img)
+center_image(firstFloor_img)
 center_image(playerF_img)
 center_image(playerB_img)
 center_image(playerL_img)
@@ -26,3 +27,4 @@ center_image(obstacle_img)
 center_image(monsterSmall_img)
 center_image(monsterMedium_img)
 center_image(monsterLarge_img)
+center_image(spawner_img)
