--- scripts/globalVar.py
@@ -1,5 +1,0 @@
-import pyglet
-
-enemy_list = []
-enemy_coord = []
-obstacle_coord = []
