--- scripts/obstacles.txt
@@ -1,98 +1,0 @@
-Layer 0
-0 0
-1 0
-2 0
-3 0
-4 0
-5 0
-6 0
-7 0
-8 0
-9 0
-10 0
-11 0
-12 0
-
-Layer 1
-0 1
-4 1
-7 1
-12 1
-
-Layer 2
-0 2
-4 2
-7 2
-12 2
-
-Layer 3
-0 3
-4 3
-5 3
-6 3
-7 3
-12 3
-
-Layer 4
-0 4
-12 4
-
-Layer 5
-0 5
-4 5
-5 5
-6 5
-7 5
-8 5
-9 5
-11 5
-12 5
-
-Layer 6
-0 6
-12 6
-
-Layer 7
-0 7
-12 7
-
-Layer 8
-0 8
-12 8
-
-Layer 9
-0 9
-12 9
-
-Layer 10
-0 10
-12 10
-
-Layer 11
-0 11
-12 11
-
-Layer 12
-0 12
-1 12
-2 12
-3 12
-4 12
-5 12
-6 12
-7 12
-8 12
-9 12
-10 12
-11 12
-12 12
-
-XXXXXX
-
-^^Reading stops when it reads this
-
-Format:
-Layer number(starting from the bottom)
-X-coord Y-coord
-
-The space between layers is 30pixels on both axes
