+++ scripts/obstacles_first.txt
@@ -1,0 +1,241 @@
+Layer 0
+8 0
+9 0
+10 0
+11 0
+12 0
+13 0
+14 0
+15 0
+16 0
+17 0
+18 0
+19 0
+20 0
+21 0
+22 0
+23 0
+
+Layer 1
+8 1
+21 1
+22 1
+23 1
+
+Layer 2
+0 2
+1 2
+2 2
+3 2
+4 2
+5 2
+6 2
+7 2
+8 2
+13 2
+14 2
+15 2 
+16 2
+17 2
+18 2
+19 2
+23 2
+
+Layer 3
+0 3
+1 3
+2 3
+8 3
+12 3
+13 3
+14 3
+15 3 
+16 3
+17 3
+18 3
+19 3
+20 3
+23 3
+
+Layer 4
+0 4
+1 4
+7 4
+8 4
+14 4
+15 4
+16 4
+20 4
+23 4
+
+Layer 5
+0 5
+1 5
+3 5
+4 5
+5 5
+7 5
+8 5
+18 5
+20 5
+22 5
+23 5
+
+Layer 6
+0 6
+1 6
+3 6
+4 6
+5 6
+7 6
+8 6
+20 6
+23 6
+
+Layer 7
+0 7
+8 7
+14 7
+15 7
+16 7
+20 7
+21 7
+23 7
+
+Layer 8
+0 8
+1 8
+2 8
+11 8
+12 8
+13 8
+14 8
+15 8
+16 8
+17 8
+20 8
+21 8
+23 8
+
+Layer 9
+0 9
+1 9
+2 9
+3 9
+4 9
+5 9
+6 9
+7 9
+8 9
+10 9
+11 9
+12 9
+13 9
+14 9
+15 9
+16 9
+17 9
+20 9
+21 9
+23 9
+
+Layer 10
+0 10
+1 10
+3 10
+4 10
+5 10
+6 10
+7 10
+10 10
+20 10
+21 10
+23 10
+
+Layer 11
+0 11
+13 11
+14 11
+15 11
+20 11
+23 11
+
+Layer 12
+0 12
+3 12
+4 12
+10 12
+13 12
+14 12
+15 12
+22 12
+23 12
+
+Layer 13
+0 13
+3 13
+4 13
+6 13
+10 13
+21 13
+22 13
+23 13
+
+Layer 14
+0 14
+1 14
+2 14
+3 14
+6 14
+4 14
+10 14
+20 14
+21 14
+22 14
+23 14
+
+Layer 15
+3 15
+4 15
+5 15
+6 15
+7 15
+8 15
+9 15
+10 15
+13 15
+14 15
+15 15
+16 15
+17 15
+20 15
+
+Layer 16
+3 16
+4 16
+5 16
+6 16
+7 16
+8 16
+9 16
+10 16
+11 16
+12 16
+13 16
+14 16
+15 16
+16 16
+17 16
+18 16
+19 16
+20 16
+
+XXXXXX
+
+^^Reading stops when it reads this
+
+Format:
+Layer number(starting from the bottom)
+X-coord Y-coord
+
+The space between layers is 30pixels on both axes
