+++ scripts/resources/enemy_spawner.png
